requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
PyMySQL
gunicorn
